requirements.txt
streamlit
python-dotenv
requests
playwright
langchain
langchain-google-genai
browser_use